Restructured project hierachy a clean makeup.
