+++ src/main/java/com/inventorymanagementsystem/ims/corestructure/LeftPaneMenuList.java
@@ -1,4 +1,4 @@
-package com.inventorymanagementsystem.ims;
+package com.inventorymanagementsystem.ims.corestructure;
 
 import java.awt.*;
 import javax.swing.JLabel;
