+++ src/main/java/com/inventorymanagementsystem/ims/corestructure/MainContentView.java
@@ -1,4 +1,4 @@
-package com.inventorymanagementsystem.ims;
+package com.inventorymanagementsystem.ims.corestructure;
 
 import javax.swing.*;
 import java.awt.*;
